Fix DisconnectListener #1 - Create component for EventListener
