+++ apps/server/src/main/java/org/gamza/server/Controller/MessageController.java
@@ -156,7 +156,6 @@ public class MessageController {
         }
 
         jsonObject.put("users", usersJsonArray);
-        System.out.println("전송되는 JSONObject : " + jsonObject);
         HttpEntity<String> request = new HttpEntity<>(jsonObject.toString(), httpHeaders);
 
         ResponseEntity<StageRequestDto> responseEntity = restTemplate.postForEntity(url, request, StageRequestDto.class);
@@ -177,68 +176,68 @@ public class MessageController {
     operations.convertAndSend("/sub/room/" + messageDto.getRoomId(), message);
   }
 
-  @EventListener
-  public void webSocketDisconnectListener(SessionDisconnectEvent event) {
-    StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
-
-    // 유저인포와 방 id 가져오기
-    UserInfo userInfo = (UserInfo) Objects.requireNonNull(accessor.getSessionAttributes()).get("userInfo");
-    Long roomId = (Long) accessor.getSessionAttributes().get("roomId");
-
-    // 유저 찾기
-    User user = userService.findByNickname(userInfo.getUser().getNickname());
-
-    // 방 찾기
-    GameRoom room = roomService.findRoom(roomId);
-
-    // 메시지 생성
-    Message message = Message.builder()
-      .type(Message.MessageType.EXIT)
-      .userInfo(UserInfo.builder().system("system").build())
-      .gameRoom(room)
-      .build();
-
-    // 방에서 해당 유저 삭제
-    roomService.removeUserToRoom(roomId, userInfo.getPlayerNumber());
-
-    // 유저 정보 수정
-    user.updateTeamStatus(TeamStatus.NONE);
-    user.updateReadyStatus(ReadyStatus.NONE);
-
-    message.setMessage(userInfo.getUser().getNickname() + "님이 퇴장하셨습니다.");
-
-    // 방이 빈 방이면 방 삭제 후 리턴
-//    if (roomService.getRoomUsers(roomId).isEmpty()) {
-//      roomRepository.delete(room);
-//      return;
+//  @EventListener
+//  public void webSocketDisconnectListener(SessionDisconnectEvent event) {
+//    StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
+//
+//    // 헤더에 저장했던 유저인포와 방 id 가져오기
+//    UserInfo userInfo = (UserInfo) Objects.requireNonNull(accessor.getSessionAttributes()).get("userInfo");
+//    Long roomId = (Long) accessor.getSessionAttributes().get("roomId");
+//
+//    // 유저 찾기
+//    User user = userService.findByNickname(userInfo.getUser().getNickname());
+//
+//    // 방 찾기
+//    GameRoom room = roomService.findRoom(roomId);
+//
+//    // 메시지 생성
+//    Message message = Message.builder()
+//      .type(Message.MessageType.EXIT)
+//      .userInfo(UserInfo.builder().system("system").build())
+//      .gameRoom(room)
+//      .build();
+//
+//    // 방에서 해당 유저 삭제
+//    roomService.removeUserToRoom(roomId, userInfo.getPlayerNumber());
+//
+//    // 유저 정보 수정
+//    user.updateTeamStatus(TeamStatus.NONE);
+//    user.updateReadyStatus(ReadyStatus.NONE);
+//
+//    message.setMessage(userInfo.getUser().getNickname() + "님이 퇴장하셨습니다.");
+//
+//    // 방이 빈 방이면 방 삭제 후 리턴
+////    if (roomService.getRoomUsers(roomId).isEmpty()) {
+////      roomRepository.delete(room);
+////      return;
+////    }
+//
+//    // 나간 애가 방장이면 방장 새로 선출
+//    if(userInfo.getUserStatus().equals(UserStatus.ROLE_MANAGER)) {
+//      selectNewHost(room);
 //    }
-
-    // 나간 애가 방장이면 방장 새로 선출
-    if(userInfo.getUserStatus().equals(UserStatus.ROLE_MANAGER)) {
-      selectNewHost(room);
-    }
-
-    operations.convertAndSend("/sub/room/" + roomId, message);
-  }
-
-  private void selectNewHost(GameRoom room) {
-    for (int i = 0; i <= room.getRoomSize(); i++) {
-      AddUserDto userDto = roomService.getRoomUsers(room.getId()).get(i);
-      User nextHost = userService.findUserByDto(userDto);
-      log.info("방장 선발");
-      if (nextHost != null) {
-        UserInfo hostInfo = UserInfo.builder()
-          .user(nextHost)
-          .userStatus(UserStatus.ROLE_MANAGER)
-          .build();
-        Message newHostMsg = Message.builder()
-          .type(Message.MessageType.ROOM)
-          .gameRoom(room)
-          .userInfo(hostInfo)
-          .message(hostInfo.getUser().getNickname() + "님이 방장이 되셨습니다.").build();
-        operations.convertAndSend("/sub/room/" + room.getId(), newHostMsg);
-        break;
-      }
-    }
-  }
+//
+//    operations.convertAndSend("/sub/room/" + roomId, message);
+//  }
+//
+//  private void selectNewHost(GameRoom room) {
+//    for (int i = 0; i <= room.getRoomSize(); i++) {
+//      AddUserDto userDto = roomService.getRoomUsers(room.getId()).get(i);
+//      User nextHost = userService.findUserByDto(userDto);
+//      log.info("방장 선발");
+//      if (nextHost != null) {
+//        UserInfo hostInfo = UserInfo.builder()
+//          .user(nextHost)
+//          .userStatus(UserStatus.ROLE_MANAGER)
+//          .build();
+//        Message newHostMsg = Message.builder()
+//          .type(Message.MessageType.ROOM)
+//          .gameRoom(room)
+//          .userInfo(hostInfo)
+//          .message(hostInfo.getUser().getNickname() + "님이 방장이 되셨습니다.").build();
+//        operations.convertAndSend("/sub/room/" + room.getId(), newHostMsg);
+//        break;
+//      }
+//    }
+//  }
 }
